Fixed Flow deployment
